state machine implemnted
I've implemented the move state machine.
I've also placed some public variables which will be used to increase
difficulty when spawning more and more drones as the game goes on.
Granted they aren't being used, but they are there to be used when the
time is right, which is not now (but soon!).